requirements.txt
gradio>=4.44.0
Pillow>=10.0.0
